sync:sync the lastest code of wander

See merge request r_d_center/department-of-ai/Nessus!97

+++ common/include/wander/detect_code.hpp
@@ -3,7 +3,6 @@
 
 #include "box_info.hpp"
 #include <abi/consumer.hpp>
-
 namespace glasssix::wander
 {
     struct detect_code;
@@ -33,7 +32,7 @@ namespace glasssix::exposing::impl
                 std::int32_t roi_y,
                 std::int32_t roi_width,
                 std::int32_t roi_height,
-                abi_in_t<exposing::param_hash_map<exposing::param_string, float>> param_map_abi,
+                abi_in_t<exposing::param_hash_map<exposing::param_string, double>> param_map_abi,
                 abi_out_t<exposing::param_vector<wander::box_info>> result) noexcept = 0;
 
             virtual std::int32_t G6_ABI_CALL version(abi_out_t<param_string> result) noexcept = 0;
@@ -62,12 +61,12 @@ namespace glasssix::exposing::impl
             std::int32_t roi_y,
             std::int32_t roi_width,
             std::int32_t roi_height,
-            abi_in_t<exposing::param_hash_map<exposing::param_string, float>> param_map_abi,
+            abi_in_t<exposing::param_hash_map<exposing::param_string, double>> param_map_abi,
             abi_out_t<exposing::param_vector<wander::box_info>> result) noexcept override
         {
             return abi_safe_call([&]
                 { *result = detach_abi(this->self().detect(create_from_abi<param_span<std::uint8_t>>(bitmap), channels, height, width, roi_x, roi_y, roi_width, roi_height,
-                   create_from_abi<exposing::param_hash_map<exposing::param_string, float>>(param_map_abi))); });
+                   create_from_abi<exposing::param_hash_map<exposing::param_string, double>>(param_map_abi))); });
         }
 
         virtual std::int32_t G6_ABI_CALL version(abi_out_t<param_string> result) noexcept override
@@ -105,7 +104,7 @@ namespace glasssix::exposing::impl
                 std::int32_t roi_y,
                 std::int32_t roi_width,
                 std::int32_t roi_height,
-                const exposing::param_hash_map<exposing::param_string, float>& param_map_abi) const
+                const exposing::param_hash_map<exposing::param_string, double>& param_map_abi) const
             {
                 exposing::param_vector<wander::box_info> result{ nullptr };
